requirements.txt
# Core Libraries
pandas==2.0.3
numpy==1.23.5

# Google Cloud SDKs
google-cloud-bigquery==3.14.1

# BigQuery Dataframe Dependencies
pyarrow>=14.0.1
db-dtypes>=1.1.1

# Authentication
oauth2client==4.1.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Environment Variables
python-dotenv==1.0.0

# CLI Progress Bar
tqdm==4.66.1

# Utility Libraries
requests==2.31.0

# Data Vizualization
matplotlib == 3.10.0